# pkg/api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Daytona Server API",
        "title": "Daytona Server API",
        "contact": {},
        "version": "v0.0.0-dev"
    },
    "host": "localhost:3986",
    "basePath": "/",
    "paths": {
...
        "/apikey/{apiKeyName}": {
            "post": {
                "description": "Generate an API key",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "apiKey"
                ],
                "summary": "Generate an API key",
                "operationId": "GenerateApiKey",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key name",
                        "name": "apiKeyName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Revoke API key",
                "tags": [
                    "apiKey"
                ],
                "summary": "Revoke API key",
                "operationId": "RevokeApiKey",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key name",
                        "name": "apiKeyName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/container-registry/{server}": {
            "get": {
                "description": "Get container registry credentials",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container-registry"
                ],
                "summary": "Get container registry credentials",
                "operationId": "GetContainerRegistry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container Registry server name",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ContainerRegistry"
                        }
                    }
                }
            },
            "put": {
                "description": "Set container registry credentials",
                "tags": [
                    "container-registry"
                ],
                "summary": "Set container registry credentials",
                "operationId": "SetContainerRegistry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container Registry server name",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Container Registry credentials to set",
                        "name": "containerRegistry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContainerRegistry"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            },
            "delete": {
                "description": "Remove a container registry credentials",
                "tags": [
                    "container-registry"
                ],
                "summary": "Remove a container registry credentials",
                "operationId": "RemoveContainerRegistry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container Registry server name",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ContainerRegistry": {
            "type": "object",
            "required": [
                "password",
                "server",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "server": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}